requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
markdown
django-filter
